Работает секьюрити

+++ src/main/java/com/example/springapp/repository/RoleRepository.java
@@ -3,7 +3,9 @@ package com.example.springapp.repository;
 import com.example.springapp.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Integer>
+import java.math.BigInteger;
+
+public interface RoleRepository extends JpaRepository<Role, BigInteger>
 {
   Role findByName(String name);
 }
